services/agent_service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0

# For agent processing
google-adk
google-genai
python-dotenv

# Local shared library  
-e ../finance_common